Add class TabularFileSink.
