+++ python/subprojects/testbed/mlrl/testbed/experiments/output/sinks/sink.py
@@ -6,7 +6,9 @@ Provides classes for implementing sinks, output data may be written to.
 from abc import ABC, abstractmethod
 from os import path
 
-from mlrl.testbed.experiments.output.data import OutputData
+from mlrl.common.config.options import Options
+
+from mlrl.testbed.experiments.output.data import OutputData, TabularOutputData
 from mlrl.testbed.experiments.state import ExperimentState
 from mlrl.testbed.util.io import get_file_name_per_fold
 
@@ -95,6 +97,40 @@ class FileSink(Sink, ABC):
     def _write_to_file(self, file_path: str, state: ExperimentState, output_data: OutputData, **kwargs):
         """
         Must be implemented by subclasses in order to write output data to a specific file.
+
+        :param file_path:   The path to the file to which the output data should be written
+        :param state:       The state from which the output data has been generated
+        :param output_data: The output data that should be written to the file
+        """
+
+
+class TabularFileSink(FileSink):
+    """
+    An abstract base class for all sinks that write tabular output data to a file.
+    """
+
+    def __init__(self, directory: str, suffix: str, options: Options = Options()):
+        """
+        :param directory:   The path to the directory of the file
+        :param suffix:      The suffix of the file
+        :param options:     Options to be taken into account
+        """
+        super().__init__(directory=directory, suffix=suffix)
+        self.options = options
+
+    def _write_to_file(self, file_path: str, state: ExperimentState, output_data: OutputData, **kwargs):
+        if not isinstance(output_data, TabularOutputData):
+            raise RuntimeError('Output data of type "' + type(output_data).__name__
+                               + '" cannot be converted into a tabular representation')
+
+        tabular_data = output_data.to_table(self.options, **kwargs)
+
+        if tabular_data:
+            self._write_table_to_file(file_path, state, tabular_data, **kwargs)
+
+    def _write_table_to_file(self, file_path: str, state: ExperimentState, table: TabularOutputData.Table, **kwargs):
+        """
+        Must be implemented by subclasses in order to write tabular output data to a specific file.
 
         :param file_path:   The path to the file to which the output data should be written
         :param state:       The state from which the output data has been generated
